S11_RAG/Summarizer/requirement.txt
langchain
langchain_openai
PyPDF2
pprintpp
python-dotenv
langchain_community